requirements.txt
# Dependencias del Sistema de Analisis de Call Center
# Instalacion: pip install -r requirements.txt

# Core ML Libraries
transformers>=4.35.0
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
datasets>=2.14.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Web Interface
gradio>=4.0.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0

# Embedding Analysis (opcional)
umap-learn>=0.5.0

# Audio Processing - REQUERIDO
openai-whisper>=20231117
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
ffmpeg-python>=0.2.0

# Desarrollo (opcional)
# jupyter>=1.0.0
# notebook>=7.0.0
# ipywidgets>=8.0.0